+++ packages/v2/voice/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/voice
 
+## 1.51.5-next.3
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.51.5-next.3
+
 ## 1.51.5-next.2
 
 ### Patch Changes